+++ pages/chat.js
@@ -1,0 +1,5 @@
+export default function ChatPage() {
+  return (
+    <div>Pagina do chat</div>
+  )
+}
